Add minor refactorings
