+++ lib/home/home_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:my_braider/home/HomeListItem.dart';
-import 'package:my_braider/home/HomeSliverAppBar.dart';
+import 'package:my_braider/home/widgets/home_list_item.dart';
+import 'package:my_braider/home/widgets/home_sliver_appbar.dart';
 import 'package:my_braider/home/model/user.dart';
 
 class HomeScreen extends StatefulWidget {
